[owners] Sync repo startup script with latest working script (#546)
* Sync repo startup script with latest working script

* Flag-guard initialization code not needed for redeployments

+++ owners/gce/startup-script.sh
@@ -20,6 +20,8 @@
 
 # [START startup]
 set -v
+#STARTUP=init
+STARTUP=deploy
 
 # [START env]
 APP_DIR="/opt/app"  # Startup
@@ -32,8 +34,45 @@ PRIVATE_KEY=$(echo | base64 -w 0 <<EOF
 -----END RSA PRIVATE KEY-----
 EOF
 )  # Probot
+# [END env]
 
-echo > "${APP_DIR}/.env" <<EOF
+supervisorctl stop nodeapp
+
+
+if [ "$STARTUP" = init ]; then
+  # [START logging]
+  # Install logging monitor. The monitor will automatically pick up logs sent to
+  # syslog.
+  curl -s "https://storage.googleapis.com/signals-agents/logging/google-fluentd-install.sh" | bash
+  service google-fluentd restart &
+  # [END logging]
+
+  # [START installs]
+  # Install dependencies from apt
+  apt-get update
+  apt-get install -yq ca-certificates git nodejs build-essential supervisor
+
+  # Install nodejs
+  mkdir /opt/nodejs
+  curl https://nodejs.org/dist/v10.15.0/node-v10.15.0-linux-x64.tar.gz | tar xvzf - -C /opt/nodejs --strip-components=1
+  ln -s /opt/nodejs/bin/node /usr/bin/node
+  ln -s /opt/nodejs/bin/npm /usr/bin/npm
+  # [END installs]
+fi
+
+
+# [START git]
+# Get the application source code from the Google Cloud Repository.
+# git requires $HOME and it's not set during the startup script.
+export HOME=/root
+git config --global credential.helper gcloud.sh
+rm -rf "${APP_DIR}"
+# TODO(#500): Deploy from `amphtml/amp-github-apps`
+git clone https://source.developers.google.com/p/amp-owners-bot/r/amp-owners-bot "${APP_DIR}"
+# [END git]
+
+# [START update]
+cat > .env <<EOF
 NODE_ENV=production
 LOG_LEVEL=trace
 PORT=8080
@@ -49,47 +88,7 @@ GITHUB_REPO_DIR=${REPO_DIR}
 GITHUB_BOT_USERNAME=amp-owners-bot
 ADD_REVIEWERS_OPT_OUT=1
 EOF
-# [END env]
 
-## Steps for: Deployment
-
-supervisorctl stop nodeapp
-
-
-## Steps for: Initialization
-
-# [START logging]
-# Install logging monitor. The monitor will automatically pick up logs sent to
-# syslog.
-curl -s "https://storage.googleapis.com/signals-agents/logging/google-fluentd-install.sh" | bash
-service google-fluentd restart &
-# [END logging]
-
-# [START installs]
-# Install dependencies from apt
-apt-get update
-apt-get install -yq ca-certificates git nodejs build-essential supervisor
-
-# Install nodejs
-mkdir /opt/nodejs
-curl https://nodejs.org/dist/v10.15.0/node-v10.15.0-linux-x64.tar.gz | tar xvzf - -C /opt/nodejs --strip-components=1
-ln -s /opt/nodejs/bin/node /usr/bin/node
-ln -s /opt/nodejs/bin/npm /usr/bin/npm
-# [END installs]
-
-# [START git]
-# Get the application source code from the Google Cloud Repository.
-# git requires $HOME and it's not set during the startup script.
-export HOME=/root
-git config --global credential.helper gcloud.sh
-rm -rf "${APP_DIR}"
-# TODO(#500): Deploy from `amphtml/amp-github-apps`
-git clone https://source.developers.google.com/p/$PROJECTID/r/amp-owners-bot "${APP_DIR}"
-# [END git]
-
-## Steps for: Deployment
-
-# [START update]
 # Install app dependencies
 cd "${APP_DIR}"
 APP_COMMIT_SHA=$(git log --max-count=1 --pretty='format:%h')  # App
@@ -98,23 +97,21 @@ npm install
 # [END update]
 
 
-## Steps for: Initialization
-# [START clone]
-# Get a clean copy of the target repository to be evaluated
-rm -rf "${REPO_DIR}"
-git clone "https://github.com/${REPO}.git" "${REPO_DIR}"
-# [END clone]
+if [ "$STARTUP" = init ]; then
+  # [START clone]
+  # Get a clean copy of the target repository to be evaluated
+  rm -rf "${REPO_DIR}"
+  git clone "https://github.com/${REPO}.git" "${REPO_DIR}"
+  # [END clone]
 
-# [START auth]
-# Create a nodeapp user. The application will run as this user.
-useradd -m -d /home/nodeapp nodeapp
-chown -R nodeapp:nodeapp "${APP_DIR}"
-chown -R nodeapp:nodeapp "${REPO_DIR}"
-# [END auth]
+  # [START auth]
+  # Create a nodeapp user. The application will run as this user.
+  useradd -m -d /home/nodeapp nodeapp
+  chown -R nodeapp:nodeapp "${REPO_DIR}"
+  # [END auth]
 
-# [START supervisor]
-# Configure supervisor to run the node app.
-cat >/etc/supervisor/conf.d/node-app.conf << EOF
+  # Configure supervisor to run the node app.
+  cat >/etc/supervisor/conf.d/node-app.conf << EOF
 [program:nodeapp]
 directory=${APP_DIR}
 command=npm run start
@@ -129,11 +126,13 @@ stderr_logfile=syslog
 stderr_logfile_maxbytes=0
 EOF
 
+fi
 
-## Steps for: Deployment
 
+# [START supervisor]
+chown -R nodeapp:nodeapp "${APP_DIR}"
 supervisorctl reread
-supervisorctl update
+supervisorctl start nodeapp
 # [END supervisor]
 
 # Application should now be running under supervisor
